int64

+++ api/models/gc.go
@@ -1,9 +1,9 @@
 package models
 
 type NormalGCPercent struct {
-	Value int `json:"value" form:"value" binding:"Required"`
+	Value int64 `json:"value" form:"value" binding:"Required"`
 }
 
 type StartupGCPercent struct {
-	Value int `json:"value" form:"value" binding:"Required"`
+	Value int64 `json:"value" form:"value" binding:"Required"`
 }
